requirements.txt
Flask==2.3.3
requests==2.31.0
confluent-kafka==2.2.0
fastavro==1.8.4
python-dotenv==0.19.0